refactor: removed react-svgbox component, and implemented it locally
There were issues with the react-svgbox component. Something with the way it was importing and/or exporting components, so it was causing errors with SSR in next.js. Removed that as a dependency, as it was really just a simple wrapper around the SVGBox API, and implemented that wrapper myself locally. Local code is nearly identical to the react-svgbox implementation, but doesn't have the same import/export issues that I was having when I had just wrapped that component.